ManagedSubject: More refinements to ManagedSubject for ease of use.
Updates ManagedSubject domain and ManagedSubjectService to suit.

ManagedSubject now specifies eduPersonAffiliation as a single valued attribute. Should we have real world need for this to be multivalued we'll address it in a future update.

ManagedSubjectService is now more friendly for controllers in the register method providing for a complete ManagedSubject to be passed in (which can be previously validated) instead of the cludgy exception method used previously.

Updated test cases to suit changes.