ManagedSubject: Finalize uses mobile number not challenge response.
In order to ensure security for VHR the AAF has decided to wear the
cost of SMS tokens and integrate SMS messaging for password recovery
instead of the previously requested Challenge Response.

Challenge Response remains in the codebase for possible future use.

These changes ask users to provide their Mobile Number as part of
finalization but do not enforce them if they don't wish to provide it.

End users not supplying a mobile number will not be able to take part
in password recovery and will need to contact their respective Group
administrators for reset.