Enhanced and extended isMutable concept.
After further business requirements came to light the work to provide
isMutable is further enhanced and extended by this commit.

We now have canCreate, canMutate and canDelete for all key objects
that take into account both actual permissions and the state of the
specific object (and its parents where needed) in determining access.

This will be folded into controllers.