ManagedSubject: Allow god admins to set login/password via CSV.
AAF determined a late requirement to provision accounts that had
pre-configured login/password and where ready for use. This was
really only required for training courses and is an edge case but none
the less needed.

CSV with 6 tokens as opposed to the normal 4 (the latter being
login,password) will activate accounts in this manner. In addition
the permission app:administrator is required.

Closes #112.