Rename 'locked' action to 'support' and change wording
This is now a generic landing page for a user whose account can not be
used to perform a password reset. As we want to disclose as little
information as possible to potential attackers, we now dump them at a
generic "contact your administrators" page with more information about
why they might not be able to reset their password.

The user ought to know whether they have finalised their account yet, so
there's very little need to customise this page per-user.