Password: Only require second factor for lost password.
As decided by AAF management the previous process on requiring users
to supply an emailed AND SMS (or administrator generated) code was
considered to complex to be understood and caused to many support
issues.

This change requires that only the second factor code is required
in order to reset a password, generally SMS.

Fallback to emailed codes is supported for environments such as test
federations which can't justify the cost associated with using SMS
for password reset.