2-Step: Store 2-Step verification for 90 days to limit input.
This commit got out of hand it should have been more broken down but
there are a lot of moving parts unfortunately.

Provides logic and test cases to ensure that users only need to
enter 2-Step verification once every 90.

Additionally this change renames the original SSO cookie to _vh_l1
to make the naming more unified with the 2-Step cookie. Being abstract
the purpose is also not quite as evident to external users.

Finally this provides minor UI changes to wording etc to make thing
easier for end users.