API: Add additional API to support basic auth for VH.
The ECP project work undertaken for NeCTAR revealed the need for the
VH to support authentication for ECP requests/sessions. This was only
possible in a practical sense using basic auth.

This change introduces an API which will validate a supplied
username/password pair for remote processes. The intedended recipient
at this time is a filter within Virtual Home.