IdP: No need for a web based login if something else set REMOTE_USER.
If the another filter or process sets REMOTE_USER and this filter is
invoked it won't attempted to redirect authentication to the VH
web interface.

This is intended to be used if the BasicAuth filter has already setup
the environment (although this filter and the BasicAuth filter should
generally have no need to operate in the same URI space).