ECP: A local commit of the ECP python test script.
This will be extended in future commits to work closely with the
ECP Validator code that has been developed. It will reflect the
core and optional attributes which are defined by the AAF and which
of those the ECP enabled service was able to obtain from VH.

The intended outcome here is very close to that of the existing
web based AAF Attribute Validator.