Conf: eduPersonEntitlement support multiple values
Fix the eduPersonEntitlement in the sample config file
configuration/idp/attribute-resolver.xml.

The original eduPersonEntitlement definition, pulls the value directly
from the database - wrong, the values need to be split in the same way
as is done for eduPersonAffiliation

Fix that by introducing a  new eduPersonEntitlement definition, based on
eduPersonAffiliation, splitting the value from the database into
multiple SAML values.  Add an extra check for a null value (when no
value had been set).

See #165 for further information.

Fixes #165.