Enhance proposed 2-Step usage of issuer
This code builds on the proposal submitted by Tuakiri for adding issuer to the generated QR code for 2-step logins.

The majority of the changes here are to bring key logic back into the Grails space, DRYing where possible and to ensure we avoid things like throwing RuntimeExceptions and printing full stack traces.

References code in Pull #181.