Refine extraction of issuer for 2-Step QR codes
As mentioned by @vladimir-mencl-eresearch the code between his commit and now incorrectly provided the issuer value for use with QR generation.

This commit ensures that double encoding is carried out refines the method name to ensure no surprise on part of the caller in getting an encoded value in response.

Test cases extended to ensure special characters are treated correctly.